+++ simpletransformers/language_modeling/language_modeling_utils.py
@@ -12,6 +12,7 @@ from tokenizers.processors import BertProcessing
 from torch.utils.data import Dataset
 from transformers import PreTrainedTokenizer
 
+
 logger = logging.getLogger(__name__)