+++ simpletransformers/t5/t5_model.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import, division, print_function
-
 import json
 import logging
 import math
@@ -7,48 +5,20 @@ import os
 import random
 import warnings
 from multiprocessing import cpu_count
+from pathlib import Path
 
 import numpy as np
-from scipy.stats import pearsonr
-from seqeval.metrics import classification_report, f1_score, precision_score, recall_score
 from tqdm.auto import tqdm, trange
 
 import pandas as pd
 import torch
 from simpletransformers.config.global_args import global_args
-from simpletransformers.ner.ner_utils import (
-    InputExample,
-    convert_examples_to_features,
-    get_examples_from_df,
-    get_labels,
-    read_examples_from_file,
-)
+from simpletransformers.t5.t5_utils import T5Dataset
 from tensorboardX import SummaryWriter
-from torch.nn import CrossEntropyLoss
-from torch.utils.data import DataLoader, RandomSampler, SequentialSampler, TensorDataset
-from transformers import (
-    WEIGHTS_NAME,
-    AdamW,
-    BertConfig,
-    BertForTokenClassification,
-    BertTokenizer,
-    CamembertConfig,
-    CamembertForTokenClassification,
-    CamembertTokenizer,
-    DistilBertConfig,
-    DistilBertForTokenClassification,
-    DistilBertTokenizer,
-    ElectraConfig,
-    ElectraForTokenClassification,
-    ElectraTokenizer,
-    RobertaConfig,
-    RobertaForTokenClassification,
-    RobertaTokenizer,
-    XLMRobertaConfig,
-    XLMRobertaForTokenClassification,
-    XLMRobertaTokenizer,
-    get_linear_schedule_with_warmup,
-)
+from torch.nn.utils.rnn import pad_sequence
+from torch.utils.data import DataLoader, Dataset, RandomSampler, SequentialSampler
+from torch.utils.data.distributed import DistributedSampler
+from transformers import AdamW, T5Config, T5ForConditionalGeneration, T5Tokenizer, get_linear_schedule_with_warmup
 
 try:
     import wandb
@@ -60,17 +30,22 @@ except ImportError:
 logger = logging.getLogger(__name__)
 
 
-class NERModel:
+def chunks(lst, n):
+    """Yield successive n-sized chunks from lst."""
+    for i in range(0, len(lst), n):
+        yield lst[i : i + n]
+
+
+class T5Model:
     def __init__(
-        self, model_type, model_name, labels=None, args=None, use_cuda=True, cuda_device=-1, **kwargs,
+        self, model_name, args=None, use_cuda=True, cuda_device=-1, **kwargs,
     ):
+
         """
-        Initializes a NERModel
+        Initializes a T5Model model.
 
         Args:
-            model_type: The type of model (bert, roberta)
-            model_name: Default Transformer model name or path to a directory containing Transformer model file (pytorch_nodel.bin).
-            labels (optional): A list of all Named Entity labels.  If not given, ["O", "B-MISC", "I-MISC",  "B-PER", "I-PER", "B-ORG", "I-ORG", "B-LOC", "I-LOC"] will be used.
+            model_name: The exact architecture and trained weights to use. This may be a Hugging Face Transformers compatible pre-trained model, a community model, or the path to a directory containing model files.
             args (optional): Default args will be used if this parameter is not provided. If provided, it should be a dict containing the args that should be changed in the default args.
             use_cuda (optional): Use GPU if available. Setting to False will force model to use CPU only.
             cuda_device (optional): Specific GPU that should be used. Will use the first available GPU by default.
@@ -84,50 +59,23 @@ class NERModel:
             if "n_gpu" in args and args["n_gpu"] > 0:
                 torch.cuda.manual_seed_all(args["manual_seed"])
 
-        if labels:
-            self.labels = labels
-        else:
-            self.labels = [
-                "O",
-                "B-MISC",
-                "I-MISC",
-                "B-PER",
-                "I-PER",
-                "B-ORG",
-                "I-ORG",
-                "B-LOC",
-                "I-LOC",
-            ]
-        self.num_labels = len(self.labels)
+        self.args = {
+            "dataset_class": None,
+            "do_sample": False,
+            "max_steps": -1,
+            "evaluate_generated_text": False,
+            "num_beams": 1,
+            "max_length": 20,
+            "repetition_penalty": 1.0,
+            "length_penalty": 2.0,
+            "early_stopping": True,
+            "preprocess_inputs": True,
+        }
 
-        self.args = {}
-        self.args = {"classification_report": False}
         self.args.update(global_args)
-
-        if not use_cuda:
-            self.args["fp16"] = False
 
         if args:
             self.args.update(args)
-
-        MODEL_CLASSES = {
-            "bert": (BertConfig, BertForTokenClassification, BertTokenizer),
-            "roberta": (RobertaConfig, RobertaForTokenClassification, RobertaTokenizer),
-            "distilbert": (DistilBertConfig, DistilBertForTokenClassification, DistilBertTokenizer),
-            "camembert": (CamembertConfig, CamembertForTokenClassification, CamembertTokenizer),
-            "xlmroberta": (XLMRobertaConfig, XLMRobertaForTokenClassification, XLMRobertaTokenizer),
-            "electra": (ElectraConfig, ElectraForTokenClassification, ElectraTokenizer),
-        }
-
-        config_class, model_class, tokenizer_class = MODEL_CLASSES[model_type]
-        if self.num_labels:
-            self.config = config_class.from_pretrained(model_name, num_labels=self.num_labels, **self.args["config"])
-            self.num_labels = self.num_labels
-        else:
-            self.config = config_class.from_pretrained(model_name, **self.args["config"])
-            self.num_labels = self.config.num_labels
-
-        self.model = model_class.from_pretrained(model_name, config=self.config, **kwargs)
 
         if use_cuda:
             if torch.cuda.is_available():
@@ -138,48 +86,46 @@ class NERModel:
             else:
                 raise ValueError(
                     "'use_cuda' set to True when cuda is unavailable."
-                    "Make sure CUDA is available or set use_cuda=False."
+                    "Make sure CUDA is available or set `use_cuda=False`."
                 )
         else:
             self.device = "cpu"
 
         self.results = {}
 
-        self.tokenizer = tokenizer_class.from_pretrained(
-            model_name, do_lower_case=self.args["do_lower_case"], **kwargs
-        )
+        self.config = T5Config.from_pretrained(model_name, **self.args["config"])
+
+        self.model = T5ForConditionalGeneration.from_pretrained(model_name, config=self.config)
+
+        self.tokenizer = T5Tokenizer.from_pretrained(model_name)
+
+        if not use_cuda:
+            self.args["fp16"] = False
 
         self.args["model_name"] = model_name
-        self.args["model_type"] = model_type
-
-        self.pad_token_label_id = CrossEntropyLoss().ignore_index
-
-        if model_type == "camembert":
-            warnings.warn(
-                "use_multiprocessing automatically disabled as CamemBERT"
-                " fails when using multiprocessing for feature conversion."
-            )
-            self.args["use_multiprocessing"] = False
 
         if self.args["wandb_project"] and not wandb_available:
             warnings.warn("wandb_project specified but wandb is not available. Wandb disabled.")
             self.args["wandb_project"] = None
 
-    def train_model(self, train_data, output_dir=None, show_running_loss=True, args=None, eval_data=None, verbose=True, **kwargs):
+    def train_model(
+        self, train_data, output_dir=None, show_running_loss=True, args=None, eval_data=None, verbose=True, **kwargs,
+    ):
         """
         Trains the model using 'train_data'
 
         Args:
-            train_data: train_data should be the path to a .txt file containing the training data OR a pandas DataFrame with 3 columns.
-                        If a text file is given the data should be in the CoNLL format. i.e. One word per line, with sentences seperated by an empty line.
-                        The first word of the line should be a word, and the last should be a Name Entity Tag.
-                        If a DataFrame is given, each sentence should be split into words, with each word assigned a tag, and with all words from the same sentence given the same sentence_id.
-            eval_data: Evaluation data (same format as train_data) against which evaluation will be performed when evaluate_during_training is enabled. Is required if evaluate_during_training is enabled.
+            train_data: Pandas DataFrame containing the 3 columns - `prefix`, `input_text`, `target_text`.
+                        - `prefix`: A string indicating the task to perform. (E.g. `"question"`, `"stsb"`)
+                        - `input_text`: The input text sequence. `prefix` is automatically prepended to form the full input. (<prefix>: <input_text>)
+                        - `target_text`: The target sequence
             output_dir: The directory where model files will be saved. If not given, self.args['output_dir'] will be used.
             show_running_loss (optional): Set to False to prevent running loss from being printed to console. Defaults to True.
             args (optional): Optional changes to the args dict of the model. Any changes made will persist for the model.
-            **kwargs: Additional metrics that should be used. Pass in the metrics as keyword arguments (name of metric: function to use). E.g. f1=sklearn.metrics.f1_score.
-                        A metric function should take in two parameters. The first parameter will be the true labels, and the second parameter will be the predictions.
+            eval_data (optional): A DataFrame against which evaluation will be performed when evaluate_during_training is enabled. Is required if evaluate_during_training is enabled.
+            **kwargs: Additional metrics that should be used. Pass in the metrics as keyword arguments (name of metric: function to use).
+                        A metric function should take in two parameters. The first parameter will be the true labels, and the second parameter will be the predictions. Both inputs
+                        will be lists of strings. Note that this will slow down training significantly as the predicted sequences need to be generated.
 
         Returns:
             None
@@ -188,21 +134,14 @@ class NERModel:
         if args:
             self.args.update(args)
 
-        if self.args["silent"]:
-            show_running_loss = False
+        # if self.args["silent"]:
+        #     show_running_loss = False
 
         if self.args["evaluate_during_training"] and eval_data is None:
-            if "eval_df" in kwargs:
-                warnings.warn(
-                    "The eval_df parameter has been renamed to eval_data."
-                    " Using eval_df will raise an error in a future version."
-                )
-                eval_data = kwargs.pop("eval_df")
-            else:
-                raise ValueError(
-                    "evaluate_during_training is enabled but eval_data is not specified."
-                    " Pass eval_data to model.train_model() if using evaluate_during_training."
-                )
+            raise ValueError(
+                "evaluate_during_training is enabled but eval_data is not specified."
+                " Pass eval_data to model.train_model() if using evaluate_during_training."
+            )
 
         if not output_dir:
             output_dir = self.args["output_dir"]
@@ -210,17 +149,22 @@ class NERModel:
         if os.path.exists(output_dir) and os.listdir(output_dir) and not self.args["overwrite_output_dir"]:
             raise ValueError(
                 "Output directory ({}) already exists and is not empty."
-                " Use --overwrite_output_dir to overcome.".format(output_dir)
+                " Set args['overwrite_output_dir'] = True to overcome.".format(output_dir)
             )
 
         self._move_model_to_device()
 
-        train_dataset = self.load_and_cache_examples(train_data)
+        train_dataset = self.load_and_cache_examples(train_data, verbose=verbose)
 
         os.makedirs(output_dir, exist_ok=True)
 
         global_step, tr_loss = self.train(
-            train_dataset, output_dir, show_running_loss=show_running_loss, eval_data=eval_data, **kwargs
+            train_dataset,
+            output_dir,
+            show_running_loss=show_running_loss,
+            eval_data=eval_data,
+            verbose=verbose,
+            **kwargs,
         )
 
         model_to_save = self.model.module if hasattr(self.model, "module") else self.model
@@ -228,24 +172,33 @@ class NERModel:
         self.tokenizer.save_pretrained(output_dir)
         torch.save(self.args, os.path.join(output_dir, "training_args.bin"))
 
-        logger.info(" Training of {} model complete. Saved to {}.".format(self.args["model_type"], output_dir))
+        if verbose:
+            logger.info(" Training of {} model complete. Saved to {}.".format(self.args["model_name"], output_dir))
 
-    def train(self, train_dataset, output_dir, show_running_loss=True, eval_data=None, verbose=True, **kwargs):
+    def train(
+        self, train_dataset, output_dir, show_running_loss=True, eval_data=None, verbose=True, **kwargs,
+    ):
         """
         Trains the model on train_dataset.
 
         Utility function to be used by the train_model() method. Not intended to be used directly.
         """
 
-        device = self.device
         model = self.model
         args = self.args
+        device = self.device
 
         tb_writer = SummaryWriter(logdir=args["tensorboard_dir"])
         train_sampler = RandomSampler(train_dataset)
         train_dataloader = DataLoader(train_dataset, sampler=train_sampler, batch_size=args["train_batch_size"])
 
-        t_total = len(train_dataloader) // args["gradient_accumulation_steps"] * args["num_train_epochs"]
+        if args["max_steps"] > 0:
+            t_total = args["max_steps"]
+            args["num_train_epochs"] = (
+                args["max_steps"] // (len(train_dataloader) // args["gradient_accumulation_steps"]) + 1
+            )
+        else:
+            t_total = len(train_dataloader) // args["gradient_accumulation_steps"] * args["num_train_epochs"]
 
         no_decay = ["bias", "LayerNorm.weight"]
         optimizer_grouped_parameters = [
@@ -253,19 +206,25 @@ class NERModel:
                 "params": [p for n, p in model.named_parameters() if not any(nd in n for nd in no_decay)],
                 "weight_decay": args["weight_decay"],
             },
-            {
-                "params": [p for n, p in model.named_parameters() if any(nd in n for nd in no_decay)],
-                "weight_decay": 0.0,
-            },
+            {"params": [p for n, p in model.named_parameters() if any(nd in n for nd in no_decay)]},
         ]
 
         warmup_steps = math.ceil(t_total * args["warmup_ratio"])
         args["warmup_steps"] = warmup_steps if args["warmup_steps"] == 0 else args["warmup_steps"]
 
-        optimizer = AdamW(optimizer_grouped_parameters, lr=args["learning_rate"], eps=args["adam_epsilon"],)
+        optimizer = AdamW(optimizer_grouped_parameters, lr=args["learning_rate"], eps=args["adam_epsilon"])
         scheduler = get_linear_schedule_with_warmup(
             optimizer, num_warmup_steps=args["warmup_steps"], num_training_steps=t_total
         )
+
+        if (
+            args["model_name"]
+            and os.path.isfile(os.path.join(args["model_name"], "optimizer.pt"))
+            and os.path.isfile(os.path.join(args["model_name"], "scheduler.pt"))
+        ):
+            # Load in optimizer and scheduler states
+            optimizer.load_state_dict(torch.load(os.path.join(args["model_name"], "optimizer.pt")))
+            scheduler.load_state_dict(torch.load(os.path.join(args["model_name"], "scheduler.pt")))
 
         if args["fp16"]:
             try:
@@ -277,6 +236,8 @@ class NERModel:
 
         if args["n_gpu"] > 1:
             model = torch.nn.DataParallel(model)
+
+        logger.info(" Training started")
 
         global_step = 0
         tr_loss, logging_loss = 0.0, 0.0
@@ -311,12 +272,13 @@ class NERModel:
 
         if args["evaluate_during_training"]:
             training_progress_scores = self._create_training_progress_scores(**kwargs)
+
         if args["wandb_project"]:
             wandb.init(project=args["wandb_project"], config={**args}, **args["wandb_kwargs"])
             wandb.watch(self.model)
 
         model.train()
-        for _ in train_iterator:
+        for current_epoch in train_iterator:
             if epochs_trained > 0:
                 epochs_trained -= 1
                 continue
@@ -328,7 +290,6 @@ class NERModel:
                 batch = tuple(t.to(device) for t in batch)
 
                 inputs = self._get_inputs_dict(batch)
-
                 outputs = model(**inputs)
                 # model outputs are always tuple in pytorch-transformers (see doc)
                 loss = outputs[0]
@@ -362,6 +323,7 @@ class NERModel:
                         torch.nn.utils.clip_grad_norm_(amp.master_params(optimizer), args["max_grad_norm"])
                     else:
                         torch.nn.utils.clip_grad_norm_(model.parameters(), args["max_grad_norm"])
+
                     optimizer.step()
                     scheduler.step()  # Update learning rate schedule
                     model.zero_grad()
@@ -370,9 +332,7 @@ class NERModel:
                     if args["logging_steps"] > 0 and global_step % args["logging_steps"] == 0:
                         # Log metrics
                         tb_writer.add_scalar("lr", scheduler.get_lr()[0], global_step)
-                        tb_writer.add_scalar(
-                            "loss", (tr_loss - logging_loss) / args["logging_steps"], global_step,
-                        )
+                        tb_writer.add_scalar("loss", (tr_loss - logging_loss) / args["logging_steps"], global_step)
                         logging_loss = tr_loss
                         if args["wandb_project"]:
                             wandb.log(
@@ -394,13 +354,16 @@ class NERModel:
                         and global_step % args["evaluate_during_training_steps"] == 0
                     ):
                         # Only evaluate when single GPU otherwise metrics may not average well
-                        results, _, _ = self.eval_model(eval_data, verbose=verbose and args["evaluate_during_training_verbose"], **kwargs)
+                        results = self.eval_model(
+                            eval_data,
+                            verbose=verbose and args["evaluate_during_training_verbose"],
+                            silent=True,
+                            **kwargs,
+                        )
                         for key, value in results.items():
                             tb_writer.add_scalar("eval_{}".format(key), value, global_step)
 
                         output_dir_current = os.path.join(output_dir, "checkpoint-{}".format(global_step))
-
-                        os.makedirs(output_dir_current, exist_ok=True)
 
                         if args["save_eval_checkpoints"]:
                             self._save_model(output_dir_current, optimizer, scheduler, model=model, results=results)
@@ -485,7 +448,9 @@ class NERModel:
                 self._save_model(output_dir_current, optimizer, scheduler, model=model)
 
             if args["evaluate_during_training"]:
-                results, _, _ = self.eval_model(eval_data, verbose=verbose and args["evaluate_during_training_verbose"], **kwargs)
+                results = self.eval_model(
+                    eval_data, verbose=verbose and args["evaluate_during_training_verbose"], silent=True, **kwargs
+                )
 
                 self._save_model(output_dir_current, optimizer, scheduler, results=results)
 
@@ -526,7 +491,6 @@ class NERModel:
                         best_eval_metric = results[args["early_stopping_metric"]]
                         self._save_model(args["best_model_dir"], optimizer, scheduler, model=model, results=results)
                         early_stopping_counter = 0
-                        early_stopping_counter = 0
                     else:
                         if args["use_early_stopping"] and args["early_stopping_consider_epochs"]:
                             if early_stopping_counter < args["early_stopping_patience"]:
@@ -549,35 +513,45 @@ class NERModel:
         Evaluates the model on eval_data. Saves results to output_dir.
 
         Args:
-            eval_data: eval_data should be the path to a .txt file containing the evaluation data or a pandas DataFrame.
-                        If a text file is used the data should be in the CoNLL format. I.e. One word per line, with sentences seperated by an empty line.
-                        The first word of the line should be a word, and the last should be a Name Entity Tag.
-                        If a DataFrame is given, each sentence should be split into words, with each word assigned a tag, and with all words from the same sentence given the same sentence_id.
-
+            eval_data: Pandas DataFrame containing the 3 columns - `prefix`, `input_text`, `target_text`.
+                        - `prefix`: A string indicating the task to perform. (E.g. `"question"`, `"stsb"`)
+                        - `input_text`: The input text sequence. `prefix` is automatically prepended to form the full input. (<prefix>: <input_text>)
+                        - `target_text`: The target sequence            
             output_dir: The directory where model files will be saved. If not given, self.args['output_dir'] will be used.
             verbose: If verbose, results will be printed to the console on completion of evaluation.
-            **kwargs: Additional metrics that should be used. Pass in the metrics as keyword arguments (name of metric: function to use). E.g. f1=sklearn.metrics.f1_score.
-                        A metric function should take in two parameters. The first parameter will be the true labels, and the second parameter will be the predictions.
-
+            silent: If silent, tqdm progress bars will be hidden.
+            **kwargs: Additional metrics that should be used. Pass in the metrics as keyword arguments (name of metric: function to use).
+                        A metric function should take in two parameters. The first parameter will be the true labels, and the second parameter will be the predictions. Both inputs
+                        will be lists of strings. Note that this will slow down evaluation significantly as the predicted sequences need to be generated.
         Returns:
-            result: Dictionary containing evaluation results. (eval_loss, precision, recall, f1_score)
-            model_outputs: List of raw model outputs
-            preds_list: List of predicted tags
+            results: Dictionary containing evaluation results.
         """  # noqa: ignore flake8"
+
         if not output_dir:
             output_dir = self.args["output_dir"]
 
         self._move_model_to_device()
 
-        eval_dataset = self.load_and_cache_examples(eval_data, evaluate=True)
+        eval_dataset = self.load_and_cache_examples(eval_data, evaluate=True, verbose=verbose, silent=silent)
+        os.makedirs(output_dir, exist_ok=True)
 
-        result, model_outputs, preds_list = self.evaluate(eval_dataset, output_dir, verbose=verbose, silent=silent, **kwargs)
+        result = self.evaluate(eval_dataset, output_dir, verbose=verbose, silent=silent, **kwargs)
         self.results.update(result)
+
+        if self.args["evaluate_generated_text"]:
+            to_predict = [
+                prefix + ": " + input_text + " </s>"
+                for prefix, input_text in zip(eval_data["prefix"], eval_data["input_text"])
+            ]
+            preds = self.predict(to_predict)
+
+            result = self.compute_metrics(eval_data["target_text"].tolist(), preds, **kwargs)
+            self.results.update(result)
 
         if verbose:
             logger.info(self.results)
 
-        return result, model_outputs, preds_list
+        return self.results
 
     def evaluate(self, eval_dataset, output_dir, verbose=True, silent=False, **kwargs):
         """
@@ -586,195 +560,130 @@ class NERModel:
         Utility function to be used by the eval_model() method. Not intended to be used directly.
         """
 
-        device = self.device
         model = self.model
         args = self.args
-        pad_token_label_id = self.pad_token_label_id
         eval_output_dir = output_dir
+        device = self.device
 
         results = {}
 
         eval_sampler = SequentialSampler(eval_dataset)
         eval_dataloader = DataLoader(eval_dataset, sampler=eval_sampler, batch_size=args["eval_batch_size"])
 
+        if args["n_gpu"] > 1:
+            model = torch.nn.DataParallel(model)
+
         eval_loss = 0.0
         nb_eval_steps = 0
-        preds = None
-        out_label_ids = None
         model.eval()
 
         for batch in tqdm(eval_dataloader, disable=args["silent"] or silent):
             batch = tuple(t.to(device) for t in batch)
 
+            inputs = self._get_inputs_dict(batch)
             with torch.no_grad():
-                inputs = {
-                    "input_ids": batch[0],
-                    "attention_mask": batch[1],
-                    "labels": batch[3],
-                }
-                # XLM and RoBERTa don"t use segment_ids
-                if args["model_type"] in ["bert", "xlnet"]:
-                    inputs["token_type_ids"] = batch[2]
                 outputs = model(**inputs)
-                tmp_eval_loss, logits = outputs[:2]
-
-                eval_loss += tmp_eval_loss.mean().item()
-
+                loss = outputs[0]
+                eval_loss += loss.mean().item()
             nb_eval_steps += 1
 
-            if preds is None:
-                preds = logits.detach().cpu().numpy()
-                out_label_ids = inputs["labels"].detach().cpu().numpy()
-            else:
-                preds = np.append(preds, logits.detach().cpu().numpy(), axis=0)
-                out_label_ids = np.append(out_label_ids, inputs["labels"].detach().cpu().numpy(), axis=0)
-
         eval_loss = eval_loss / nb_eval_steps
-        model_outputs = preds
-        preds = np.argmax(preds, axis=2)
 
-        label_map = {i: label for i, label in enumerate(self.labels)}
-
-        out_label_list = [[] for _ in range(out_label_ids.shape[0])]
-        preds_list = [[] for _ in range(out_label_ids.shape[0])]
-
-        for i in range(out_label_ids.shape[0]):
-            for j in range(out_label_ids.shape[1]):
-                if out_label_ids[i, j] != pad_token_label_id:
-                    out_label_list[i].append(label_map[out_label_ids[i][j]])
-                    preds_list[i].append(label_map[preds[i][j]])
-
-        extra_metrics = {}
-        for metric, func in kwargs.items():
-            extra_metrics[metric] = func(out_label_list, preds_list)
-
-        result = {
-            "eval_loss": eval_loss,
-            "precision": precision_score(out_label_list, preds_list),
-            "recall": recall_score(out_label_list, preds_list),
-            "f1_score": f1_score(out_label_list, preds_list),
-            **extra_metrics
-        }
-
-        results.update(result)
+        results["eval_loss"] = eval_loss
 
         output_eval_file = os.path.join(eval_output_dir, "eval_results.txt")
         with open(output_eval_file, "w") as writer:
-            if args["classification_report"]:
-                cls_report = classification_report(out_label_list, preds_list)
-                writer.write("{}\n".format(cls_report))
-            for key in sorted(result.keys()):
-                writer.write("{} = {}\n".format(key, str(result[key])))
+            for key in sorted(results.keys()):
+                writer.write("{} = {}\n".format(key, str(results[key])))
 
-        return results, model_outputs, preds_list
+        return results
 
-    def predict(self, to_predict, split_on_space=True):
+    def predict(self, to_predict):
         """
         Performs predictions on a list of text.
 
         Args:
-            to_predict: A python list of text (str) to be sent to the model for prediction.
-            split_on_space: If True, each sequence will be split by spaces for assigning labels.
-                            If False, to_predict must be a a list of lists, with the inner list being a
-                            list of strings consisting of the split sequences. The outer list is the list of sequences to
-                            predict on.
+            to_predict: A python list of text (str) to be sent to the model for prediction. Note that the prefix should be prepended to the text.
 
         Returns:
-            preds: A Python list of lists with dicts containing each word mapped to its NER tag.
-            model_outputs: A python list of the raw model outputs for each text.
-        """
-
-        device = self.device
-        model = self.model
-        args = self.args
-        pad_token_label_id = self.pad_token_label_id
+            preds: A python list of the generated sequences.
+        """  # noqa: ignore flake8"
 
         self._move_model_to_device()
 
-        if split_on_space:
-            predict_examples = [
-                InputExample(i, sentence.split(), [self.labels[0] for word in sentence.split()])
-                for i, sentence in enumerate(to_predict)
-            ]
-        else:
-            predict_examples = [
-                InputExample(i, sentence, [self.labels[0] for word in sentence])
-                for i, sentence in enumerate(to_predict)
-            ]
-
-        eval_dataset = self.load_and_cache_examples(None, to_predict=predict_examples)
-
-        eval_sampler = SequentialSampler(eval_dataset)
-        eval_dataloader = DataLoader(eval_dataset, sampler=eval_sampler, batch_size=args["eval_batch_size"])
-
-        eval_loss = 0.0
-        nb_eval_steps = 0
-        preds = None
-        out_label_ids = None
-        model.eval()
-
-        for batch in tqdm(eval_dataloader, disable=args["silent"]):
-            batch = tuple(t.to(device) for t in batch)
-
-            with torch.no_grad():
-                inputs = {
-                    "input_ids": batch[0],
-                    "attention_mask": batch[1],
-                    "labels": batch[3],
-                }
-                # XLM and RoBERTa don"t use segment_ids
-                if args["model_type"] in ["bert", "xlnet"]:
-                    inputs["token_type_ids"] = batch[2]
-                outputs = model(**inputs)
-                tmp_eval_loss, logits = outputs[:2]
-
-                eval_loss += tmp_eval_loss.mean().item()
-
-            nb_eval_steps += 1
-
-            if preds is None:
-                preds = logits.detach().cpu().numpy()
-                out_label_ids = inputs["labels"].detach().cpu().numpy()
+        all_outputs = []
+        # Batching
+        for batch in [
+            to_predict[i : i + self.args["eval_batch_size"]]
+            for i in range(0, len(to_predict), self.args["eval_batch_size"])
+        ]:
+            if self.args["preprocess_inputs"]:
+                input_ids = self.tokenizer.batch_encode_plus(
+                    [t + " </s>" for t in batch],
+                    max_length=self.args["max_seq_length"],
+                    pad_to_max_length=True,
+                    return_tensors="pt",
+                )["input_ids"]
             else:
-                preds = np.append(preds, logits.detach().cpu().numpy(), axis=0)
-                out_label_ids = np.append(out_label_ids, inputs["labels"].detach().cpu().numpy(), axis=0)
+                input_ids = self.tokenizer.batch_encode_plus(
+                    batch, max_length=self.args["max_seq_length"], pad_to_max_length=True, return_tensors="pt",
+                )["input_ids"]
+            input_ids = input_ids.to(self.device)
+            outputs = self.model.generate(
+                input_ids=input_ids,
+                num_beams=self.args["num_beams"],
+                max_length=self.args["max_length"],
+                length_penalty=self.args["length_penalty"],
+                early_stopping=self.args["early_stopping"],
+                repetition_penalty=self.args["repetition_penalty"],
+                do_sample=self.args["do_sample"],
+            )
+            all_outputs.extend(outputs)
 
-        eval_loss = eval_loss / nb_eval_steps
-        model_outputs = preds
-        preds = np.argmax(preds, axis=2)
-
-        label_map = {i: label for i, label in enumerate(self.labels)}
-
-        out_label_list = [[] for _ in range(out_label_ids.shape[0])]
-        preds_list = [[] for _ in range(out_label_ids.shape[0])]
-
-        for i in range(out_label_ids.shape[0]):
-            for j in range(out_label_ids.shape[1]):
-                if out_label_ids[i, j] != pad_token_label_id:
-                    out_label_list[i].append(label_map[out_label_ids[i][j]])
-                    preds_list[i].append(label_map[preds[i][j]])
-
-        preds = [
-            [{word: preds_list[i][j]} for j, word in enumerate(sentence.split()[: len(preds_list[i])])]
-            for i, sentence in enumerate(to_predict)
+        return [
+            self.tokenizer.decode(output_id, skip_special_tokens=True, clean_up_tokenization_spaces=True)
+            for output_id in all_outputs
         ]
 
-        return preds, model_outputs
-
-    def load_and_cache_examples(self, data, evaluate=False, no_cache=False, to_predict=None):
+    def compute_metrics(self, labels, preds, **kwargs):
         """
-        Reads data_file and generates a TensorDataset containing InputFeatures. Caches the InputFeatures.
-        Utility function for train() and eval() methods. Not intended to be used directly.
+        Computes the evaluation metrics for the model predictions.
 
         Args:
-            data: Path to a .txt file containing training or evaluation data OR a pandas DataFrame containing 3 columns - sentence_id, words, labels.
-                    If a DataFrame is given, each sentence should be split into words, with each word assigned a tag, and with all words from the same sentence given the same sentence_id.
-            evaluate (optional): Indicates whether the examples are for evaluation or for training.
-            no_cache (optional): Force feature conversion and prevent caching. I.e. Ignore cached features even if present.
+            labels: List of target sequences
+            preds: List of model generated outputs
+            **kwargs: Custom metrics that should be used. Pass in the metrics as keyword arguments (name of metric: function to use).
+                        A metric function should take in two parameters. The first parameter will be the true labels, and the second parameter will be the predictions. Both inputs
+                        will be lists of strings. Note that this will slow down evaluation significantly as the predicted sequences need to be generated.
 
+        Returns:
+            result: Dictionary containing evaluation results.
         """  # noqa: ignore flake8"
+        assert len(labels) == len(preds)
 
-        process_count = self.args["process_count"]
+        results = {}
+        for metric, func in kwargs.items():
+            results[metric] = func(labels, preds)
+
+        return results
+
+    def _move_model_to_device(self):
+        self.model.to(self.device)
+
+    def _get_inputs_dict(self, batch):
+        lm_labels = batch[1]
+        lm_labels[lm_labels == self.tokenizer.pad_token_id] = -100
+
+        inputs = {"input_ids": batch[0], "lm_labels": batch[1]}
+
+        return inputs
+
+    def load_and_cache_examples(self, data, evaluate=False, no_cache=False, verbose=True, silent=False):
+        """
+        Creates a T5Dataset from data.
+
+        Utility function for train() and eval() methods. Not intended to be used directly.
+        """
 
         tokenizer = self.tokenizer
         args = self.args
@@ -782,100 +691,31 @@ class NERModel:
         if not no_cache:
             no_cache = args["no_cache"]
 
+        os.makedirs(self.args["cache_dir"], exist_ok=True)
+
         mode = "dev" if evaluate else "train"
 
-        if not to_predict:
-            if isinstance(data, str):
-                examples = read_examples_from_file(data, mode)
-            else:
-                examples = get_examples_from_df(data)
+        if args["dataset_class"]:
+            CustomDataset = args["dataset_class"]
+            return CustomDataset(tokenizer, args, data, mode)
         else:
-            examples = to_predict
-            no_cache = True
+            return T5Dataset(tokenizer, self.args, data, mode,)
 
-        cached_features_file = os.path.join(
-            args["cache_dir"],
-            "cached_{}_{}_{}_{}_{}".format(
-                mode, args["model_type"], args["max_seq_length"], self.num_labels, len(examples),
-            ),
-        )
-        if not no_cache:
-            os.makedirs(self.args["cache_dir"], exist_ok=True)
-
-        if os.path.exists(cached_features_file) and (
-            (not args["reprocess_input_data"] and not no_cache)
-            or (mode == "dev" and args["use_cached_eval_features"] and not no_cache)
-        ):
-            features = torch.load(cached_features_file)
-            logger.info(f" Features loaded from cache at {cached_features_file}")
-        else:
-            logger.info(f" Converting to features started.")
-            features = convert_examples_to_features(
-                examples,
-                self.labels,
-                self.args["max_seq_length"],
-                self.tokenizer,
-                # XLNet has a CLS token at the end
-                cls_token_at_end=bool(args["model_type"] in ["xlnet"]),
-                cls_token=tokenizer.cls_token,
-                cls_token_segment_id=2 if args["model_type"] in ["xlnet"] else 0,
-                sep_token=tokenizer.sep_token,
-                # RoBERTa uses an extra separator b/w pairs of sentences,
-                # cf. github.com/pytorch/fairseq/commit/1684e166e3da03f5b600dbb7855cb98ddfcd0805
-                sep_token_extra=bool(args["model_type"] in ["roberta"]),
-                # PAD on the left for XLNet
-                pad_on_left=bool(args["model_type"] in ["xlnet"]),
-                pad_token=tokenizer.convert_tokens_to_ids([tokenizer.pad_token])[0],
-                pad_token_segment_id=4 if args["model_type"] in ["xlnet"] else 0,
-                pad_token_label_id=self.pad_token_label_id,
-                process_count=process_count,
-                silent=args["silent"],
-                use_multiprocessing=args["use_multiprocessing"],
-            )
-
-            if not no_cache:
-                torch.save(features, cached_features_file)
-
-        all_input_ids = torch.tensor([f.input_ids for f in features], dtype=torch.long)
-        all_input_mask = torch.tensor([f.input_mask for f in features], dtype=torch.long)
-        all_segment_ids = torch.tensor([f.segment_ids for f in features], dtype=torch.long)
-        all_label_ids = torch.tensor([f.label_ids for f in features], dtype=torch.long)
-
-        dataset = TensorDataset(all_input_ids, all_input_mask, all_segment_ids, all_label_ids)
-
-        return dataset
-
-    def _move_model_to_device(self):
-        self.model.to(self.device)
-
-    def _get_last_metrics(self, metric_values):
-        return {metric: values[-1] for metric, values in metric_values.items()}
-
-    def _get_inputs_dict(self, batch):
-        inputs = {
-            "input_ids": batch[0],
-            "attention_mask": batch[1],
-            "labels": batch[3],
-        }
-        # XLM, DistilBERT, RoBERTa, and XLM-RoBERTa don't use segment_ids
-        if self.args["model_type"] in ["bert", "xlnet", "albert"]:
-            inputs["token_type_ids"] = batch[2]
-
-        return inputs
+        return T5Dataset
 
     def _create_training_progress_scores(self, **kwargs):
         extra_metrics = {key: [] for key in kwargs}
         training_progress_scores = {
             "global_step": [],
-            "precision": [],
-            "recall": [],
-            "f1_score": [],
-            "train_loss": [],
             "eval_loss": [],
+            "train_loss": [],
             **extra_metrics,
         }
 
         return training_progress_scores
+
+    def _get_last_metrics(self, metric_values):
+        return {metric: values[-1] for metric, values in metric_values.items()}
 
     def _save_model(self, output_dir, optimizer, scheduler, model=None, results=None):
         os.makedirs(output_dir, exist_ok=True)